+++ nf_core/modules/modules_repo.py
@@ -18,11 +18,47 @@ class ModulesRepo(object):
 
     def __init__(self, repo="nf-core/modules", branch="master"):
         self.name = repo
-        self.owner, self.repo = self.name.split("/")
         self.branch = branch
+
+        # Verify that the repo seems to be correctly configured
+        if self.name != "nf-core/modules" or self.branch != "master":
+            try:
+                self.verify_modules_repo()
+            except LookupError:
+                raise
+
+        self.owner, self.repo = self.name.split("/")
         self.modules_file_tree = {}
         self.modules_current_hash = None
         self.modules_avail_module_names = []
+
+    def verify_modules_repo(self):
+
+        # Check if name seems to be well formed
+        if self.name.count("/") != 1:
+            raise LookupError(f"Repository name '{self.name}' is not well formed")
+
+        # Check if repository exist
+        api_url = f"https://api.github.com/repos/{self.name}/branches"
+        response = requests.get(api_url)
+        if response.status_code == 200:
+            branches = [branch["name"] for branch in response.json()]
+            if self.branch not in branches:
+                raise LookupError(f"Branch '{self.branch}' not found in '{self.name}'")
+        else:
+            raise LookupError(f"Repository '{self.name}' is not available from GitHub")
+
+        api_url = f"https://api.github.com/repos/{self.name}/contents?ref={self.branch}"
+        response = requests.get(api_url)
+        if response.status_code == 200:
+            dir_names = [entry["name"] for entry in response.json() if entry["type"] == "dir"]
+            if "modules" not in dir_names:
+                err_str = f"Repository '{self.name}' ({self.branch}) does not contain a 'modules' directory"
+                if "software" in dir_names:
+                    err_str += ".\nAs of version 2.0, the 'software' directory should be renamed to 'modules'"
+                raise LookupError(err_str)
+        else:
+            raise LookupError(f"Unable to fetch repository information from '{self.name}' ({self.branch})")
 
     def get_modules_file_tree(self):
         """
@@ -50,6 +86,8 @@ class ModulesRepo(object):
             if f["path"].startswith(f"modules/") and f["path"].endswith("/main.nf") and "/test/" not in f["path"]:
                 # remove modules/ and /main.nf
                 self.modules_avail_module_names.append(f["path"].replace("modules/", "").replace("/main.nf", ""))
+        if len(self.modules_avail_module_names) == 0:
+            raise LookupError(f"Found no modules in '{self.name}'")
 
     def get_module_file_urls(self, module, commit=""):
         """Fetch list of URLs for a specific module