+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 * Added `--all` flag to `nf-core modules install`
 * Added `remote` and `local` subcommands to `nf-core modules list`
 * Fix bug due to restructuring in modules template
+* Added checks for verifying that the remote repository is well formed
 
 #### Sync